# app/static/swagger.yaml
swagger: "2.0"
info:
  title: "Mechanic API"
  description: "This API is used to create, update, and delete information for a mechanic shop"
  version: "1.0.0"
host: "127.0.0.1:5000"
schemes:
  - http
consumes:
  - application/json
produces:
  - application/json
securityDefinitions:
  bearerAuth:
    type: apiKey
    name: Authorization
    in: header
    description: |
      Token Required!
      Please enter your JWT token **with `Bearer` prefix** like:
      ```
      Bearer "JWT TOKEN"
      ```
paths:
...
  /employees/login:
    post:
      tags:
        - Employees
      summary: "login endpoint for employees, (generates token)"
      description: "Validates employee credentials, generates and return a employee specific token"
      consumes:
        - application/json
      produces:
        - application/json
      parameters:
        - in: body
          name: body
          description: Login Credentials for employee's
          required: true
          schema:
            $ref: "#/definitions/EmployeeLoginCredentials"
      responses:
        200:
          description: Successful login
          schema:
            $ref: "#/definitions/EmployeeLoginResponse"
        401:
          description: Unauthorized - Invalid credentials
          schema:
            $ref: "#/definitions/ErrorResponse"
  /employees:
    post:
      tags:
        - Employees
      summary: "Endpoint to create a new employee."
      description: "Endpoint to create a new employee in our mechanic shop system. emails must be unique"
      parameters:
        - in: "body"
          name: "body"
          required: true
          schema:
            $ref: "#/definitions/CreateEmployeePayload"
      responses:
        201:
          description: "New Employee added successfully"
          schema:
            $ref: "#/definitions/CreateEmployeeResponse"
          examples:
            application/json:
              id: 1
              name: John Doe
              email: JDoe@example.com
              phone: 111-222-3333
              address: 123 Main St
              role: mechanic
              salary: 100000
        404:
          description: "Validation Error - Missing or invalid fields"
          schema:
            $ref: "#/definitions/ErrorResponse"

    get:
      tags:
        - Employees
      summary: "Return all employee's"
      description: "Returns a paginated list of all employee's"
      parameters:
        - name: page
          in: query
          type: integer
          required: false
          description: "Page number (default: 1)"
        - name: per_page
          in: query
          type: integer
          required: false
          description: "Number of results per page (default: 10)"
      responses:
        200:
          description: "Retrieved Employees Successfully"
          schema:
            $ref: "#/definitions/AllEmployees"
          examples:
            application/json:
              employees:
                - id: 2
                  name: Joe
                  email: Joe@example.com
                  phone: 111-222-3333
                  address: 123 Main St
                  password: "******"
                  role: mechanic
                  salary: 100000
                - id: 3
                  name: Jane
                  email: Jane@example.com
                  phone: 222-333-4444
                  address: 123 Main St
                  password: "******"
                  role: mechanic
                  salary: 100000
              page: 1
              per_page: 10
              total_employees: 3
  /employees/working_tickets:
    get:
      tags:
        - Employees
      summary: "Returns all tickets employees are working on"
      description: "Returns an array of tickets being worked on by employee's"
      produces:
        - application/json
      responses:
        200:
          description: "Successfully retrieved ticket data for mechanics"
          schema:
            type: array
            items:
              type: object
              properties:
                id:
                  type: integer
                  description: "Mechanic's ID"
                name:
                  type: string
                  description: "Mechanic's name"
                ticket_count:
                  type: integer
                  description: "Number of tickets the mechanic is currently working on"
                ticket_ids:
                  type: array
                  items:
                    type: integer
                  description: "List of ticket IDs assigned to the mechanic"
          examples:
            application/json:
              - id: 1
                name: "John Smith"
                ticket_count: 3
                ticket_ids: [101, 102, 103]
              - id: 2
                name: "Alice Johnson"
                ticket_count: 2
                ticket_ids: [104, 105]
        500:
          description: "Internal Server Error"
          schema:
            $ref: "#/definitions/ErrorResponse"
  /employees/{id}:
    get:
      tags:
        - Employees
      summary: "Retrieve a single employee by ID"
      parameters:
        - name: id
          in: path
          required: true
          type: integer
          description: "ID of the employee to retrieve"
      responses:
        200:
          description: "Employee retrieved successfully"
          schema:
            $ref: "#/definitions/CreateEmployeeResponse"
        404:
          description: Employee not found
          schema:
            $ref: "#/definitions/ErrorResponse"

    put:
      tags:
        - Employees
      summary: "Endpoint to update a employee"
      description: "Endpoint to update a employee, this is a token authenticated route, and the user must be logged in to update their account"
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          type: integer
          description: "ID of the employee to update"
        - in: body
          name: body
          description: "Information required to update a Employee"
          required: true
          schema:
            $ref: "#/definitions/UpdateEmployeePayload"
      responses:
        200:
          description: "Successfully Updated Employee"
          schema:
            $ref: "#/definitions/UpdateEmployeeResponse"
          examples:
            application/json:
              id: 1
              name: "Rob Down"
              email: "dobr@example.com"
              phone: "111-222-3333"
              role: "mechanic"
              salary: 1000000
        401:
          description: Unauthorized - Missing or Invalid token
          schema:
            $ref: "#/definitions/ErrorResponse"
        403:
          description: Forbidden - Not allowed to perform this action
          schema:
            $ref: "#/definitions/ErrorResponse"
        404:
          description: Employee not found
          schema:
            $ref: "#/definitions/ErrorResponse"

    delete:
      tags:
        - Employees
      summary: "Endpoint to delete an Employee"
      description: "Endpoint to delete a employee. Requires token authentication, employee must be logged in to delete their account"
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          type: integer
          description: "ID of the employee to delete"
      responses:
        200:
          description: "Successfully Deleted Employee"
          schema:
            $ref: "#/definitions/DeleteEmployeeResponse"
          examples:
            application/json:
              message: "Successfully deleted employee <employee_id>"
        401:
          description: Unauthorized - Missing or invalid token
          schema:
            $ref: "#/definitions/ErrorResponse"
        403:
          description: Forbidden - Not allowed to perform this action
          schema:
            $ref: "#/definitions/ErrorResponse"
        404:
          description: Employee not found
          schema:
            $ref: "#/definitions/ErrorResponse"
...
definitions:
...
  ErrorResponse:
    type: "object"
    properties:
      message:
        type: "string"
...
  EmployeeLoginCredentials:
    type: "object"
    required:
      - email
      - password
      - role
    properties:
      email:
        type: "string"
        description: "Employee's email address"
      password:
        type: "string"
        description: "Employee's password"
      role:
        type: "string"
        description: "Role of the employee gives certain permissions to different actions in the system (e.g. creating a ticket)"
  EmployeeLoginResponse:
    type: "object"
    properties:
      status:
        type: "string"
        description: "Status of the login attempt"
      auth_token:
        type: "string"
        description: "JWT access token"
      message:
        type: "string"
        description: "Response message"
  CreateEmployeePayload:
    type: "object"
    required:
      - name
      - email
      - phone
      - address
      - password
      - role
      - salary
    properties:
      name:
        type: "string"
        description: "Employee's Full name"
      email:
        type: "string"
        description: "Employee's email"
      phone:
        type: "string"
        description: "Employee's phone number (e.g. 111-222-3333)"
      address:
        type: "string"
        description: "Employee's home address"
      password:
        type: "string"
        description: "Employee-created password"
      role:
        type: "string"
        description: "identifies the user as an employee"
      salary:
        type: "integer"
        description: "How much the employee make's"
  CreateEmployeeResponse:
    type: "object"
    properties:
      id:
        type: "integer"
      name:
        type: "string"
      email:
        type: "string"
      phone:
        type: "string"
      password:
        type: "string"
      address:
        type: "string"
      role:
        type: "string"
      salary:
        type: "integer"
  AllEmployees:
    type: "object"
    properties:
      employees:
        type: "array"
        items:
          $ref: "#/definitions/CreateEmployeeResponse"
      page:
        type: "integer"
      per_page:
        type: "integer"
      total_employees:
        type: "integer"
  UpdateEmployeePayload:
    allOf:
      - $ref: "#/definitions/CreateEmployeeResponse"
  UpdateEmployeeResponse:
    type: "object"
    properties:
      id:
        type: "integer"
      employee:
        $ref: "#/definitions/CreateEmployeeResponse"
  DeleteEmployeeResponse:
    type: "object"
    properties:
      message:
        type: "string"
...